requirements.txt
Markdown==2.2.1
PyMeta==0.5.0
PyYAML==3.10
RethinkORM==0.2.0
WebOb==1.2.3
WebTest==2.0.1
Whoosh==2.4.1
arrow==0.3.4
beautifulsoup4==4.1.3
bleach==1.2.1
coverage==3.6
docopt==0.6.1
gevent==0.13.8
gevent-fastcgi==0.1.15dev
gevent-zeromq==0.2.5
greenlet==0.4.0
hiredis==0.1.1
html5lib==0.95
itsdangerous==0.21
nose==1.3.0
protobuf==2.5.0
py-bcrypt==0.2
pystache==0.5.3
python-dateutil==2.1
pytz==2012j
pyzmq==2.2.0
redis==2.7.2
requests==1.2.3
rethinkdb==1.7.0-2
six==1.2.0
waitress==0.8.2
wsgiref==0.1.2
zope.interface==4.0.4
